treewide: unquote EAPIs
I figured we should standardize one way or the other, with or without
quotes, and I don't want to think of the magick required to add quotes
to every EAPI assignment in each ebuild. This should ease using git grep
to find ebuilds with outdated EAPIs as we go along.